+++ src/main/kotlin/app/domain/CurrencyRate.kt
@@ -1,10 +1,12 @@
 package app.domain
 
+import java.math.BigDecimal
+
 /**
  * Currency rate.
  *
  * @param rateToBase rate to convert to base
  * @param rateFromBase rate to convert from base
  */
-class CurrencyRate(val rateToBase: Double,
-                   val rateFromBase: Double)
+class CurrencyRate(val rateToBase: BigDecimal,
+                   val rateFromBase: BigDecimal)
